Commit Changes in Revert Deposit Payment

+++ app/DepositPayment.php
@@ -47,7 +47,7 @@ class DepositPayment extends Model
 
     public function revertData($user_id){
         
-        $data['amount'] = (int) $this->amount;
+        $data['amount'] = (double) $this->amount;
         $data['transaction_number'] = 'RD'.str_replace('.','',microtime(true));
         $data['payment_method_id'] = (int) $this->payment_method_id;
         $data['notes'] = 'Revertion for ' . $this->transaction_number;
